refactor: Utility functions for form object
